Rename components back to .jsx and update imports
