+++ src/components/DailyCheckIn.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Card } from './ui/card.js';
 import { Button } from './ui/button.js';
 import { Textarea } from './ui/textarea.js';
-import SectionTitle from './SectionTitle.js';
+import SectionTitle from './SectionTitle.jsx';
 import { useCollection } from '../firebase.js';
 
 export default function DailyCheckIn({ userId }) {
